did some server testing
